feat: remove move modal context
